Rename frequenz.sdk.configs and move the actor
The module frequenz.sdk.configs is renamed to frequenz.sdk.config (no
"s") and the `ConfigManager` actor is moved to
`frequenz.sdk.actor.ConfigManagingActor`.
