+++ src/frequenz/sdk/actor/__init__.py
@@ -4,6 +4,7 @@
 """A base class for creating simple composable actors."""
 
 from ._channel_registry import ChannelRegistry
+from ._config_managing import ConfigManagingActor
 from ._data_sourcing import ComponentMetricRequest, DataSourcingActor
 from ._decorator import actor
 from ._resampling import ComponentMetricsResamplingActor
@@ -12,6 +13,7 @@ __all__ = [
     "ChannelRegistry",
     "ComponentMetricRequest",
     "ComponentMetricsResamplingActor",
+    "ConfigManagingActor",
     "DataSourcingActor",
     "actor",
 ]
