+++ src/frequenz/sdk/data_ingestion/microgrid_data.py
@@ -19,7 +19,7 @@ from frequenz.channels import Broadcast, Receiver, Sender
 from frequenz.channels.util import Merge, Select
 
 from ..actor import actor
-from ..configs import Config
+from ..config import Config
 from ..data_handling.time_series import TimeSeriesEntry
 from ..microgrid.client import MicrogridApiClient
 from ..microgrid.graph import ComponentGraph
